+++ test/src/test_bit_field.cpp
@@ -13,17 +13,17 @@ TEST(test_bit_field, bit_field_read_as)
 {
 
     auto field = bitter::bit_field<uint8_t, 1U>(1U);
-    EXPECT_TRUE(field.read_as<bool>());
+    EXPECT_TRUE(field.as<bool>());
 }
 
 TEST(test_bit_field, bit_field_read_as1)
 {
     auto field = bitter::bit_field<uint8_t, 1U>(0U);
-    EXPECT_FALSE(field.read_as<bool>());
+    EXPECT_FALSE(field.as<bool>());
 }
 
 TEST(test_bit_field, bit_field_read_as2)
 {
     auto field = bitter::bit_field<uint8_t, 8U>(255U);
-    EXPECT_EQ(255U, field.read_as<uint8_t>());
+    EXPECT_EQ(255U, field.as<uint8_t>());
 }