+++ test/src/test_readme.cpp
@@ -46,10 +46,10 @@ TEST(test_readme, reading_a_lsb0_bit_field)
 {
     auto reader = bitter::lsb0_reader<bitter::u32, 8, 8, 8, 8>(0x12345678);
 
-    uint8_t value0 = reader.field<0>().read_as<uint8_t>(); // Read bits 0-7
-    uint8_t value1 = reader.field<1>().read_as<uint8_t>(); // Read bits 8-15
-    uint8_t value2 = reader.field<2>().read_as<uint8_t>(); // Read bits 16-23
-    uint8_t value3 = reader.field<3>().read_as<uint8_t>(); // Read bits 24-31
+    uint8_t value0 = reader.field<0>().as<uint8_t>(); // Read bits 0-7
+    uint8_t value1 = reader.field<1>().as<uint8_t>(); // Read bits 8-15
+    uint8_t value2 = reader.field<2>().as<uint8_t>(); // Read bits 16-23
+    uint8_t value3 = reader.field<3>().as<uint8_t>(); // Read bits 24-31
 
     assert(value0 == 0x78);
     assert(value1 == 0x56);
@@ -61,10 +61,10 @@ TEST(test_readme, reading_a_msb0_bit_field)
 {
     auto reader = bitter::msb0_reader<bitter::u32, 8, 8, 8, 8>(0x12345678);
 
-    uint8_t value0 = reader.field<0>().read_as<uint8_t>(); // Read bits 0-7
-    uint8_t value1 = reader.field<1>().read_as<uint8_t>(); // Read bits 8-15
-    uint8_t value2 = reader.field<2>().read_as<uint8_t>(); // Read bits 16-23
-    uint8_t value3 = reader.field<3>().read_as<uint8_t>(); // Read bits 24-31
+    uint8_t value0 = reader.field<0>().as<uint8_t>(); // Read bits 0-7
+    uint8_t value1 = reader.field<1>().as<uint8_t>(); // Read bits 8-15
+    uint8_t value2 = reader.field<2>().as<uint8_t>(); // Read bits 16-23
+    uint8_t value3 = reader.field<3>().as<uint8_t>(); // Read bits 24-31
 
     assert(value0 == 0x12);
     assert(value1 == 0x34);